Add BG/FG registry support to colortool
colortool apparently only supports setting the fg / bg index for the currently running console. This commit includes an experimental version of colortool that supports setting the BF/FG index to the registry so it's used as the default for all consoles (unless the console shortcut overrides the defaults in the registry, via a subkey).