add uninstall support
This is a bit tricky, because Windows stores color configuration in both a hierarchical registry format, as well as in each shortcut (.lnk) file.

This should cover the "default setup" for most people, but it won't handle the cases where users have their own shortcuts set up. I don't think there's a way to handle this last scenario except for searching the hard drive for all .lnk files, which is out-of-scope for this project.